+++ CHANGELOG.md
@@ -36,4 +36,5 @@
   - Context hook
   - Config hook
   - Export as system properties
+  - Un-export system properties
   - Launcher invocation (conditional)